fix Viewer references and add Instance convenience methods
The annoying code pattern `Instance(ls_instances(name="my_instance")[0])`
is overly common. Adds a method `Instance.find`, which is essentially
shorthand for this, but raises a (suppressible) exception if more than
one instance is found. It is similar in intent to `Cluster.from_descriptions`
and may be further aliased or distinguished in the future.

Also adds the method `Instance.make_ssh_string`, which returns a string
that, if run locally in a system shell, might, possibly, maybe, making
lots of assumptions about system configuration, start an interactive
SSH session on the instance. This is intended to avoid copy-paste
or window-switching annoyances. It should not be taken to suggest that
this command has anything to do with `Instance`'s inner workings.

Also, the `__getattr_` method of `Viewer`s following remote processes
 in some cases inappropriately attempted to access methods of an
object abstracting the underlying OS-level process (which does not
exist for remote processes), producing confusing and irrelevant error
messages. Such cases now result in normal AttributeErrors.