some light refactoring and signature changes to Bucket
the ability to pass ExtraArgs/extra_args to additional methods of Bucket
has become important. Permit passing them in a more 'natural'
way as varkwargs. This could plausibly break some code that used a
bunch of positional arguments to these methods, but I think that the
chance this is an issue is smaller than the benefit of

Eventually we may want to change a lot of arguments to Bucket's high-level
methods to keyword-only. This will also make dispatch to multiple
threads / processes less cumbersome.